k8s-agent-graph/requirements.txt
langgraph>=0.0.20
langchain>=0.1.0
openai>=1.12.0
python-dotenv>=1.0.0
rich>=13.7.0
click>=8.1.7
pydantic>=2.6.1 